+++ ent/parser/parser.go
@@ -44,11 +44,11 @@ func (p *Engine) PreprocessAndParse(
 	keepHTML bool,
 	capitalize bool,
 	preserveDiaereses bool,
-	noSpacedInitials bool,
+	compactAuthors bool,
 ) ScientificNameNode {
 	p.code = code
 	p.preserveDiaereses = preserveDiaereses
-	p.noSpacedInitials = noSpacedInitials
+	p.compactAuthors = compactAuthors
 
 	originalString := s
 	var tagsOrEntities, lowCase bool