+++ interface.go
@@ -19,8 +19,8 @@ type GNparser interface {
 	// Debug parses a string and outputs raw AST tree from PEG engine.
 	Debug(s string) []byte
 
-	// FlatOutput returns whether flatten output is enabled.
-	FlatOutput() bool
+	// WithFlatOutput returns whether flatten output is enabled.
+	WithFlatOutput() bool
 
 	// Format returns currently chosen desired output format of a JSON or
 	// CSV output.
@@ -43,7 +43,4 @@ type GNparser interface {
 	// ParseNames takes a slice of name-strings, and returns a slice of
 	// parsed results in the same order as the input.
 	ParseNames([]string) []parsed.Parsed
-
-	// WebLogs returns a boolean to show or not the web-service logs.
-	WebLogs() bool
 }